Refactored code to move notification trigger to its own middleware
