+++ packages/notifications-provider/package.json
@@ -8,7 +8,9 @@
   },
   "author": "jm.gilgado@gmail.com",
   "dependencies": {
-    "@bufferapp/async-data-fetch": "1.9.4"
+    "@bufferapp/async-data-fetch": "1.9.4",
+    "@bufferapp/notifications": "1.9.4",
+    "@bufferapp/publish-initial-loading": "2.0.0"
   },
   "devDependencies": {
     "eslint": "3.19.0"
