+++ packages/store/index.js
@@ -29,6 +29,7 @@ import { middleware as dateTimePreferencesMiddleware } from '@bufferapp/date-tim
 import { middleware as closeAccountMiddleware } from '@bufferapp/close-account';
 import { middleware as maintenanceRedirectMiddleware } from '@bufferapp/maintenance-redirect';
 import { middleware as defaultPageMiddleware } from '@bufferapp/default-page';
+import { middleware as notificationsProviderMiddleware } from '@bufferapp/publish-notifications-provider';
 // Remove datePickerMiddleware when publish switches to analyze
 import { middleware as datePickerMiddleware } from '@bufferapp/analyze-date-picker';
 import performanceMiddleware from '@bufferapp/performance-tracking/middleware';
@@ -86,6 +87,7 @@ const configureStore = (initialstate) => {
         bufferMetricsMiddleware,
         draftsMiddleware,
         datePickerMiddleware,
+        notificationsProviderMiddleware,
       ),
     ),
   );
